slideserver/requirements.txt
# slideserver/requirements.txt
Flask==3.1.1
openslide-python==1.4.2
Pillow==11.2.1
gunicorn==23.0.0